requirements.txt
colorama~=0.4.6
pytz~=2024.1
pandas~=2.2.2
db-dtypes~=1.3.1
numpy~=2.2.1
google-api-core~=2.24.0
google-auth~=2.37.0
google-cloud-bigquery~=3.27.0
google-cloud-bigquery-storage~=2.27.0
google-cloud-storage~=2.19.0
google-api-python-client~=2.156.0
google-cloud-secret-manager~=2.22.0
google-cloud-language~=2.16.0
protobuf~=5.29.2
requests~=2.32.3
pyspark==3.5.2
findspark==1.4.2